course: refactor add student component
this just moves all the code for the add student component to its own
component and file instead of being tangled up with the student list

motivation:
 - cleaner code
 - we can now also make a popup modal to add a student that is actived by a menu
   item
